requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
binance-connector==1.5.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
constantly==15.1.0
cryptography==3.4.8
dateparser==1.0.0
Flask==2.0.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==19.0.0
PySocks==1.7.1
python-binance==1.0.12
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
tweepy==3.10.0
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
websockets==9.1
Werkzeug==2.0.1
yarl==1.6.3
zope.interface==5.4.0
